Platform compile flexibility, more readable logic
This initial commit does two things. One is it makes a few tweaks to allow the .exe to correctly compile so it will run on more platforms. This is only tested on my computer but my computer is a platform it wasn't going to compile on before so hey! The included .ico file is necessary (or another .ico if you want different icons I guess!), and now all the various file scan functions just look for any of /r/n, /n, or /r for line breaks instead of asking your particular platform what it thinks a newline is because different platforms have different opinions on this and if their opinion before was not \n you were going to be in for a rough time. Note that still this github repository does not include external dependencies; this is just to make the core .exe for the randomizer.

I added the labels to the REQ_CASUAL file. I spotted several bits of logic I thought might be erroneous but I changed none of them; I want to work with more knowledgeable MM players on logic upgrades. These labels, just taken from the user logic format, should make this file human readable which is important to working with it. The location names are still pretty difficult to understand in a lot of cases if you don't know everything about MM, and the ordering isn't great from a human perspective (things are often not next to other things in the same area), but fixing that up is a project for another time.